Module Storage_functors.Make_data_set_storage

Parameters

module C : Raw_context.T

module I : INDEX

Signature

type t = C.t

type context = t

type elt = I.t

The type of elements.

val mem : context -> elt -> bool Lwt.t

Tells if a elt is a member of the set

val add : context -> elt -> Raw_context.t Lwt.t

Adds a elt is a member of the set

val del : context -> elt -> Raw_context.t Lwt.t

Removes a elt of the set ; does nothing if not a member

val set : context -> elt -> bool -> Raw_context.t Lwt.t

Adds/Removes a elt of the set

val elements : context -> elt list Lwt.t

Returns the elements of the set, deserialized in a list in no particular order.

val fold : context -> init:'a -> f:(elt -> 'a -> 'a Lwt.t) -> 'a Lwt.t

Iterates over the elements of the set.

val clear : context -> Raw_context.t Lwt.t

Removes all elements in the set